Changed names of EFEOverlapOperator class to PristineEFEOverlapOperator and OrthoEFEOverlapOperator classes
