+++ src/basis/CMakeLists.txt
@@ -21,7 +21,8 @@ set(DFT-EFE-BASIS-SOURCES
 	AtomIdsPartition.cpp
 	EnrichmentIdsPartition.cpp
 	CFEOverlapOperatorContext.cpp
-	EFEOverlapOperatorContext.cpp
+	PristineEFEOverlapOperatorContext.cpp
+	OrthoEFEOverlapOperatorContext.cpp
 	EnrichmentClassicalInterfaceSpherical.cpp
 	L2ProjectionLinearSolverFunction.cpp
 	Defaults.cpp
