+++ src/basis/PristineEFEOverlapOperatorContext.cpp
@@ -23,4 +23,4 @@
  * @author Avirup Sircar
  */
 
-#include <basis/EFEOverlapOperatorContext.h>
+#include <basis/PristineEFEOverlapOperatorContext.h>
