+++ src/basis/PristineEFEOverlapOperatorContext.t.cpp
@@ -35,7 +35,7 @@ namespace dftefe
 {
   namespace basis
   {
-    namespace EFEOverlapOperatorContextInternal
+    namespace PristineEFEOverlapOperatorContextInternal
     {
       template <typename ValueTypeOperator,
                 typename ValueTypeOperand,
@@ -63,7 +63,7 @@ namespace dftefe
         utils::throwException(
           feBDH != nullptr,
           "Could not cast BasisDofHandler to EFEBasisDofHandler "
-          "in EFEOverlapOperatorContext");
+          "in PristineEFEOverlapOperatorContext");
 
         const size_type numLocallyOwnedCells = feBDH->nLocallyOwnedCells();
         dofsInCellVec.resize(numLocallyOwnedCells, 0);
@@ -170,7 +170,7 @@ namespace dftefe
         utils::throwException(
           cfeBDH != nullptr,
           "Could not cast BasisDofHandler to FEBasisDofHandler "
-          "in EFEOverlapOperatorContext");
+          "in PristineEFEOverlapOperatorContext");
 
         std::shared_ptr<const EFEBasisDofHandler<ValueTypeOperand,
                                                  ValueTypeOperator,
@@ -185,7 +185,7 @@ namespace dftefe
         utils::throwException(
           efeBDH != nullptr,
           "Could not cast BasisDofHandler to EFEBasisDofHandler "
-          "in EFEOverlapOperatorContext");
+          "in PristineEFEOverlapOperatorContext");
 
         // NOTE: cellId 0 passed as we assume only H refined in this function
 
@@ -319,457 +319,6 @@ namespace dftefe
         utils::MemoryTransfer<memorySpace, utils::MemorySpace::HOST>::copy(
           basisOverlapTmp.size(), basisOverlap->data(), basisOverlapTmp.data());
       }
-
-      // Use this for data storage of orthogonalized EFE only
-      template <typename ValueTypeOperator,
-                typename ValueTypeOperand,
-                utils::MemorySpace memorySpace,
-                size_type          dim>
-      void
-      computeBasisOverlapMatrix(
-        const FEBasisDataStorage<ValueTypeOperator, memorySpace>
-          &classicalBlockBasisDataStorage,
-        const FEBasisDataStorage<ValueTypeOperator, memorySpace>
-          &enrichmentBlockEnrichmentBasisDataStorage,
-        const FEBasisDataStorage<ValueTypeOperator, memorySpace>
-          &enrichmentBlockClassicalBasisDataStorage,
-        std::shared_ptr<utils::MemoryStorage<ValueTypeOperator, memorySpace>>
-          &                     basisOverlap,
-        std::vector<size_type> &cellStartIdsBasisOverlap,
-        std::vector<size_type> &dofsInCellVec)
-      {
-        std::shared_ptr<
-          const FEBasisDofHandler<ValueTypeOperand, memorySpace, dim>>
-          ccfeBDH = std::dynamic_pointer_cast<
-            const FEBasisDofHandler<ValueTypeOperand, memorySpace, dim>>(
-            classicalBlockBasisDataStorage.getBasisDofHandler());
-        utils::throwException(
-          ccfeBDH != nullptr,
-          "Could not cast BasisDofHandler to FEBasisDofHandler "
-          "in EFEOverlapOperatorContext for the Classical data storage of classical dof block.");
-
-        std::shared_ptr<
-          const FEBasisDofHandler<ValueTypeOperand, memorySpace, dim>>
-          ecfeBDH = std::dynamic_pointer_cast<
-            const FEBasisDofHandler<ValueTypeOperand, memorySpace, dim>>(
-            enrichmentBlockClassicalBasisDataStorage.getBasisDofHandler());
-        utils::throwException(
-          ecfeBDH != nullptr,
-          "Could not cast BasisDofHandler to FEBasisDofHandler "
-          "in EFEOverlapOperatorContext for the Classical data storage of enrichment dof blocks.");
-
-        std::shared_ptr<const EFEBasisDofHandler<ValueTypeOperand,
-                                                 ValueTypeOperator,
-                                                 memorySpace,
-                                                 dim>>
-          eefeBDH = std::dynamic_pointer_cast<
-            const EFEBasisDofHandler<ValueTypeOperand,
-                                     ValueTypeOperator,
-                                     memorySpace,
-                                     dim>>(
-            enrichmentBlockEnrichmentBasisDataStorage.getBasisDofHandler());
-        utils::throwException(
-          eefeBDH != nullptr,
-          "Could not cast BasisDofHandler to EFEBasisDofHandler "
-          "in EFEOverlapOperatorContext for the Enrichment data storage of enrichment dof blocks.");
-
-        utils::throwException(
-          ccfeBDH->getTriangulation() == ecfeBDH->getTriangulation() &&
-            ccfeBDH->getFEOrder(0) == ecfeBDH->getFEOrder(0) &&
-            ccfeBDH->getTriangulation() == eefeBDH->getTriangulation() &&
-            ccfeBDH->getFEOrder(0) == eefeBDH->getFEOrder(0),
-          "The EFEBasisDataStorage and and Classical FEBasisDataStorage have different triangulation or FEOrder"
-          "in EFEOverlapOperatorContext.");
-
-        utils::throwException(
-          eefeBDH->isOrthogonalized(),
-          "The Enrcihment data storage of enrichment dof blocks should have isOrthogonalized as true in EFEOverlapOperatorContext.");
-
-        std::shared_ptr<
-          const EnrichmentClassicalInterfaceSpherical<ValueTypeOperator,
-                                                      memorySpace,
-                                                      dim>>
-          eci = eefeBDH->getEnrichmentClassicalInterface();
-
-        size_type nTotalEnrichmentIds =
-          eci->getEnrichmentIdsPartition()->nTotalEnrichmentIds();
-
-        // interpolate the ci 's to the enrichment quadRuleAttr quadpoints
-
-        const EFEBasisDataStorage<ValueTypeOperator, memorySpace>
-          &enrichmentBlockEnrichmentBasisDataStorageEFE = dynamic_cast<
-            const EFEBasisDataStorage<ValueTypeOperator, memorySpace> &>(
-            enrichmentBlockEnrichmentBasisDataStorage);
-        utils::throwException(
-          &enrichmentBlockEnrichmentBasisDataStorageEFE != nullptr,
-          "Could not cast FEBasisDataStorage to EFEBasisDataStorage "
-          "in EFEOverlapOperatorContext for enrichmentBlockEnrichmentBasisDataStorage.");
-
-        // Set up the overlap matrix quadrature storages.
-
-        const size_type numLocallyOwnedCells = eefeBDH->nLocallyOwnedCells();
-        dofsInCellVec.resize(numLocallyOwnedCells, 0);
-        cellStartIdsBasisOverlap.resize(numLocallyOwnedCells, 0);
-        size_type cumulativeBasisOverlapId = 0;
-
-        size_type       basisOverlapSize = 0;
-        size_type       cellId           = 0;
-        const size_type feOrder          = eefeBDH->getFEOrder(cellId);
-
-        size_type       dofsPerCell;
-        const size_type dofsPerCellCFE = ccfeBDH->nCellDofs(cellId);
-
-        bool isConstantDofsAndQuadPointsInCellClassicalBlock           = false,
-             isConstantDofsAndQuadPointsInCellEnrichmentBlockClassical = false;
-        quadrature::QuadratureFamily quadFamilyClassicalBlock =
-          classicalBlockBasisDataStorage.getQuadratureRuleContainer()
-            ->getQuadratureRuleAttributes()
-            .getQuadratureFamily();
-        quadrature::QuadratureFamily quadFamilyEnrichmentBlockClassical =
-          enrichmentBlockClassicalBasisDataStorage.getQuadratureRuleContainer()
-            ->getQuadratureRuleAttributes()
-            .getQuadratureFamily();
-        if ((quadFamilyClassicalBlock == quadrature::QuadratureFamily::GAUSS ||
-             quadFamilyClassicalBlock == quadrature::QuadratureFamily::GLL) &&
-            !ccfeBDH->isVariableDofsPerCell())
-          isConstantDofsAndQuadPointsInCellClassicalBlock = true;
-        if ((quadFamilyEnrichmentBlockClassical ==
-               quadrature::QuadratureFamily::GAUSS ||
-             quadFamilyEnrichmentBlockClassical ==
-               quadrature::QuadratureFamily::GLL) &&
-            !ecfeBDH->isVariableDofsPerCell())
-          isConstantDofsAndQuadPointsInCellEnrichmentBlockClassical = true;
-
-        auto locallyOwnedCellIter = eefeBDH->beginLocallyOwnedCells();
-
-        for (; locallyOwnedCellIter != eefeBDH->endLocallyOwnedCells();
-             ++locallyOwnedCellIter)
-          {
-            dofsInCellVec[cellId] = eefeBDH->nCellDofs(cellId);
-            basisOverlapSize += dofsInCellVec[cellId] * dofsInCellVec[cellId];
-            cellId++;
-          }
-
-        std::vector<ValueTypeOperator> basisOverlapTmp(0);
-
-        basisOverlap = std::make_shared<
-          utils::MemoryStorage<ValueTypeOperator, memorySpace>>(
-          basisOverlapSize);
-        basisOverlapTmp.resize(basisOverlapSize, ValueTypeOperator(0));
-
-        auto      basisOverlapTmpIter = basisOverlapTmp.begin();
-        size_type cellIndex           = 0;
-
-        const utils::MemoryStorage<ValueTypeOperator, memorySpace>
-          &basisDataInAllCellsClassicalBlock =
-            classicalBlockBasisDataStorage.getBasisDataInAllCells();
-        const utils::MemoryStorage<ValueTypeOperator, memorySpace>
-          &basisDataInAllCellsEnrichmentBlockClassical =
-            enrichmentBlockClassicalBasisDataStorage.getBasisDataInAllCells();
-        const utils::MemoryStorage<ValueTypeOperator, memorySpace>
-          &basisDataInAllCellsEnrichmentBlockEnrichment =
-            enrichmentBlockEnrichmentBasisDataStorage.getBasisDataInAllCells();
-
-        size_type cumulativeClassicalDofQuadPointsOffset                 = 0,
-                  cumulativeEnrichmentBlockClassicalDofQuadPointsOffset  = 0,
-                  cumulativeEnrichmentBlockEnrichmentDofQuadPointsOffset = 0;
-
-        //
-        const std::unordered_map<global_size_type,
-                                 utils::OptimizedIndexSet<size_type>>
-          *enrichmentIdToClassicalLocalIdMap =
-            &eci->getClassicalComponentLocalIdsMap();
-        const std::unordered_map<global_size_type,
-                                 std::vector<ValueTypeOperator>>
-          *enrichmentIdToInterfaceCoeffMap =
-            &eci->getClassicalComponentCoeffMap();
-        std::shared_ptr<const FEBasisManager<ValueTypeOperator,
-                                             ValueTypeOperator,
-                                             memorySpace,
-                                             dim>>
-          cfeBasisManager =
-            std::dynamic_pointer_cast<const FEBasisManager<ValueTypeOperator,
-                                                           ValueTypeOperator,
-                                                           memorySpace,
-                                                           dim>>(
-              eci->getCFEBasisManager());
-
-        locallyOwnedCellIter = eefeBDH->beginLocallyOwnedCells();
-        for (; locallyOwnedCellIter != eefeBDH->endLocallyOwnedCells();
-             ++locallyOwnedCellIter)
-          {
-            dofsPerCell = dofsInCellVec[cellIndex];
-            size_type nQuadPointInCellClassicalBlock =
-              classicalBlockBasisDataStorage.getQuadratureRuleContainer()
-                ->nCellQuadraturePoints(cellIndex);
-            std::vector<double> cellJxWValuesClassicalBlock =
-              classicalBlockBasisDataStorage.getQuadratureRuleContainer()
-                ->getCellJxW(cellIndex);
-
-            size_type nQuadPointInCellEnrichmentBlockClassical =
-              enrichmentBlockClassicalBasisDataStorage
-                .getQuadratureRuleContainer()
-                ->nCellQuadraturePoints(cellIndex);
-            std::vector<double> cellJxWValuesEnrichmentBlockClassical =
-              enrichmentBlockClassicalBasisDataStorage
-                .getQuadratureRuleContainer()
-                ->getCellJxW(cellIndex);
-
-            size_type nQuadPointInCellEnrichmentBlockEnrichment =
-              enrichmentBlockEnrichmentBasisDataStorage
-                .getQuadratureRuleContainer()
-                ->nCellQuadraturePoints(cellIndex);
-            std::vector<double> cellJxWValuesEnrichmentBlockEnrichment =
-              enrichmentBlockEnrichmentBasisDataStorage
-                .getQuadratureRuleContainer()
-                ->getCellJxW(cellIndex);
-
-            const ValueTypeOperator *cumulativeClassicalBlockDofQuadPoints =
-              basisDataInAllCellsClassicalBlock.data() +
-              cumulativeClassicalDofQuadPointsOffset;
-
-            const ValueTypeOperator
-              *cumulativeEnrichmentBlockClassicalDofQuadPoints =
-                basisDataInAllCellsEnrichmentBlockClassical.data() +
-                cumulativeEnrichmentBlockClassicalDofQuadPointsOffset;
-
-            const ValueTypeOperator
-              *cumulativeEnrichmentBlockEnrichmentDofQuadPoints =
-                basisDataInAllCellsEnrichmentBlockEnrichment.data() +
-                cumulativeEnrichmentBlockEnrichmentDofQuadPointsOffset;
-
-            std::vector<utils::Point> quadRealPointsVec =
-              enrichmentBlockEnrichmentBasisDataStorage
-                .getQuadratureRuleContainer()
-                ->getCellRealPoints(cellIndex);
-
-            std::vector<size_type> vecClassicalLocalNodeId(0);
-
-            size_type numEnrichmentIdsInCell = dofsPerCell - dofsPerCellCFE;
-
-            std::vector<ValueTypeOperator> classicalComponentInQuadValuesEC(0);
-
-            classicalComponentInQuadValuesEC.resize(
-              nQuadPointInCellEnrichmentBlockClassical * numEnrichmentIdsInCell,
-              (ValueTypeOperator)0);
-
-            if (numEnrichmentIdsInCell > 0)
-              {
-                cfeBasisManager->getCellDofsLocalIds(cellIndex,
-                                                     vecClassicalLocalNodeId);
-
-                std::vector<ValueTypeOperator> coeffsInCell(
-                  dofsPerCellCFE * numEnrichmentIdsInCell, 0);
-
-                for (size_type cellEnrichId = 0;
-                     cellEnrichId < numEnrichmentIdsInCell;
-                     cellEnrichId++)
-                  {
-                    // get the enrichmentIds
-                    global_size_type enrichmentId =
-                      eci->getEnrichmentId(cellIndex, cellEnrichId);
-
-                    // get the vectors of non-zero localIds and coeffs
-
-                    auto iter =
-                      enrichmentIdToInterfaceCoeffMap->find(enrichmentId);
-                    DFTEFE_Assert(iter !=
-                                  enrichmentIdToInterfaceCoeffMap->end());
-                    const std::vector<ValueTypeOperator> &coeffsInLocalIdsMap =
-                      iter->second;
-
-                    for (size_type i = 0; i < dofsPerCellCFE; i++)
-                      {
-                        size_type pos   = 0;
-                        bool      found = false;
-                        auto      it =
-                          enrichmentIdToClassicalLocalIdMap->find(enrichmentId);
-                        DFTEFE_Assert(it !=
-                                      enrichmentIdToClassicalLocalIdMap->end());
-                        it->second.getPosition(vecClassicalLocalNodeId[i],
-                                               pos,
-                                               found);
-                        if (found)
-                          {
-                            coeffsInCell[numEnrichmentIdsInCell * i +
-                                         cellEnrichId] =
-                              coeffsInLocalIdsMap[pos];
-                          }
-                      }
-                  }
-
-                dftefe::utils::MemoryStorage<ValueTypeOperator,
-                                             utils::MemorySpace::HOST>
-                  basisValInCellEC =
-                    enrichmentBlockClassicalBasisDataStorage.getBasisDataInCell(
-                      cellIndex);
-
-                // Do a gemm (\Sigma c_i N_i^classical)
-                // and get the quad values in std::vector
-
-                linearAlgebra::blasLapack::gemm<ValueTypeOperator,
-                                                ValueTypeOperator,
-                                                utils::MemorySpace::HOST>(
-                  linearAlgebra::blasLapack::Layout::ColMajor,
-                  linearAlgebra::blasLapack::Op::NoTrans,
-                  linearAlgebra::blasLapack::Op::Trans,
-                  numEnrichmentIdsInCell,
-                  nQuadPointInCellEnrichmentBlockClassical,
-                  dofsPerCellCFE,
-                  (ValueTypeOperator)1.0,
-                  coeffsInCell.data(),
-                  numEnrichmentIdsInCell,
-                  basisValInCellEC.data(),
-                  nQuadPointInCellEnrichmentBlockClassical,
-                  (ValueTypeOperator)0.0,
-                  classicalComponentInQuadValuesEC.data(),
-                  numEnrichmentIdsInCell,
-                  *eci->getLinAlgOpContext());
-              }
-
-            for (unsigned int iNode = 0; iNode < dofsPerCell; iNode++)
-              {
-                for (unsigned int jNode = 0; jNode < dofsPerCell; jNode++)
-                  {
-                    *basisOverlapTmpIter = 0.0;
-                    // Ni_classical* Ni_classical of the classicalBlockBasisData
-                    if (iNode < dofsPerCellCFE && jNode < dofsPerCellCFE)
-                      {
-                        for (unsigned int qPoint = 0;
-                             qPoint < nQuadPointInCellClassicalBlock;
-                             qPoint++)
-                          {
-                            *basisOverlapTmpIter +=
-                              *(cumulativeClassicalBlockDofQuadPoints +
-                                nQuadPointInCellClassicalBlock * iNode +
-                                qPoint) *
-                              *(cumulativeClassicalBlockDofQuadPoints +
-                                nQuadPointInCellClassicalBlock * jNode +
-                                qPoint) *
-                              cellJxWValuesClassicalBlock[qPoint];
-                          }
-                      }
-
-                    else if (iNode >= dofsPerCellCFE && jNode < dofsPerCellCFE)
-                      {
-                        ValueTypeOperator NpiNcj = (ValueTypeOperator)0,
-                                          NciNcj = (ValueTypeOperator)0;
-                        // Ni_pristine*Ni_classical at quadpoints
-                        for (unsigned int qPoint = 0;
-                             qPoint < nQuadPointInCellEnrichmentBlockEnrichment;
-                             qPoint++)
-                          {
-                            NpiNcj +=
-                              eefeBDH->getEnrichmentValue(
-                                cellIndex,
-                                iNode - dofsPerCellCFE,
-                                quadRealPointsVec[qPoint]) *
-                              *(cumulativeEnrichmentBlockEnrichmentDofQuadPoints +
-                                nQuadPointInCellEnrichmentBlockEnrichment *
-                                  jNode +
-                                qPoint) *
-                              cellJxWValuesEnrichmentBlockEnrichment[qPoint];
-                          }
-
-                        // Ni_classical using Mc = d quadrature * interpolated
-                        // ci's in Ni_classicalQuadrature of Mc = d
-                        for (unsigned int qPoint = 0;
-                             qPoint < nQuadPointInCellEnrichmentBlockClassical;
-                             qPoint++)
-                          {
-                            NciNcj +=
-                              classicalComponentInQuadValuesEC
-                                [numEnrichmentIdsInCell * qPoint +
-                                 (iNode - dofsPerCellCFE)] *
-                              *(cumulativeEnrichmentBlockClassicalDofQuadPoints +
-                                nQuadPointInCellEnrichmentBlockClassical *
-                                  jNode +
-                                qPoint) *
-                              cellJxWValuesEnrichmentBlockClassical[qPoint];
-                          }
-                        *basisOverlapTmpIter += NpiNcj - NciNcj;
-                      }
-
-                    else if (iNode < dofsPerCellCFE && jNode >= dofsPerCellCFE)
-                      {
-                        ValueTypeOperator NciNpj = (ValueTypeOperator)0,
-                                          NciNcj = (ValueTypeOperator)0;
-                        // Ni_pristine*Ni_classical at quadpoints
-                        for (unsigned int qPoint = 0;
-                             qPoint < nQuadPointInCellEnrichmentBlockEnrichment;
-                             qPoint++)
-                          {
-                            NciNpj +=
-                              *(cumulativeEnrichmentBlockEnrichmentDofQuadPoints +
-                                nQuadPointInCellEnrichmentBlockEnrichment *
-                                  iNode +
-                                qPoint) *
-                              eefeBDH->getEnrichmentValue(
-                                cellIndex,
-                                jNode - dofsPerCellCFE,
-                                quadRealPointsVec[qPoint]) *
-                              cellJxWValuesEnrichmentBlockEnrichment[qPoint];
-                          }
-
-                        // Ni_classical using Mc = d quadrature * interpolated
-                        // ci's in Ni_classicalQuadrature of Mc = d
-                        for (unsigned int qPoint = 0;
-                             qPoint < nQuadPointInCellEnrichmentBlockClassical;
-                             qPoint++)
-                          {
-                            NciNcj +=
-                              *(cumulativeEnrichmentBlockClassicalDofQuadPoints +
-                                nQuadPointInCellEnrichmentBlockClassical *
-                                  iNode +
-                                qPoint) *
-                              classicalComponentInQuadValuesEC
-                                [numEnrichmentIdsInCell * qPoint +
-                                 (jNode - dofsPerCellCFE)] *
-                              cellJxWValuesEnrichmentBlockClassical[qPoint];
-                          }
-                        *basisOverlapTmpIter += NciNpj - NciNcj;
-                      }
-
-                    else if (iNode >= dofsPerCellCFE && jNode >= dofsPerCellCFE)
-                      {
-                        // Ni_pristine*Ni_pristine at quadpoints
-                        for (unsigned int qPoint = 0;
-                             qPoint < nQuadPointInCellEnrichmentBlockEnrichment;
-                             qPoint++)
-                          {
-                            *basisOverlapTmpIter +=
-                              *(cumulativeEnrichmentBlockEnrichmentDofQuadPoints +
-                                nQuadPointInCellEnrichmentBlockEnrichment *
-                                  iNode +
-                                qPoint) *
-                              *(cumulativeEnrichmentBlockEnrichmentDofQuadPoints +
-                                nQuadPointInCellEnrichmentBlockEnrichment *
-                                  jNode +
-                                qPoint) *
-                              cellJxWValuesEnrichmentBlockEnrichment[qPoint];
-                          }
-                      }
-                    basisOverlapTmpIter++;
-                  }
-              }
-
-            cellStartIdsBasisOverlap[cellIndex] = cumulativeBasisOverlapId;
-            cumulativeBasisOverlapId += dofsPerCell * dofsPerCell;
-            cellIndex++;
-            if (!isConstantDofsAndQuadPointsInCellClassicalBlock)
-              cumulativeClassicalDofQuadPointsOffset +=
-                nQuadPointInCellClassicalBlock * dofsPerCellCFE;
-            if (!isConstantDofsAndQuadPointsInCellEnrichmentBlockClassical)
-              cumulativeEnrichmentBlockClassicalDofQuadPointsOffset +=
-                nQuadPointInCellEnrichmentBlockClassical * dofsPerCellCFE;
-            cumulativeEnrichmentBlockEnrichmentDofQuadPointsOffset +=
-              nQuadPointInCellEnrichmentBlockEnrichment * dofsPerCell;
-          }
-
-        utils::MemoryTransfer<memorySpace, utils::MemorySpace::HOST>::copy(
-          basisOverlapTmp.size(), basisOverlap->data(), basisOverlapTmp.data());
-      }
-
 
       template <utils::MemorySpace memorySpace>
       void
@@ -919,7 +468,7 @@ namespace dftefe
             utils::MemoryStorage<size_type, memorySpace> strideC(
               numCellsInBlock);
 
-            EFEOverlapOperatorContextInternal::storeSizes(
+            PristineEFEOverlapOperatorContextInternal::storeSizes(
               mSizes,
               nSizes,
               kSizes,
@@ -996,18 +545,18 @@ namespace dftefe
           }
       }
 
-    } // end of namespace EFEOverlapOperatorContextInternal
+    } // end of namespace PristineEFEOverlapOperatorContextInternal
 
 
     template <typename ValueTypeOperator,
               typename ValueTypeOperand,
               utils::MemorySpace memorySpace,
               size_type          dim>
-    EFEOverlapOperatorContext<ValueTypeOperator,
-                              ValueTypeOperand,
-                              memorySpace,
-                              dim>::
-      EFEOverlapOperatorContext(
+    PristineEFEOverlapOperatorContext<ValueTypeOperator,
+                                      ValueTypeOperand,
+                                      memorySpace,
+                                      dim>::
+      PristineEFEOverlapOperatorContext(
         const FEBasisManager<ValueTypeOperand,
                              ValueTypeOperator,
                              memorySpace,
@@ -1031,7 +580,7 @@ namespace dftefe
 
       std::shared_ptr<utils::MemoryStorage<ValueTypeOperator, memorySpace>>
         basisOverlap;
-      EFEOverlapOperatorContextInternal::computeBasisOverlapMatrix<
+      PristineEFEOverlapOperatorContextInternal::computeBasisOverlapMatrix<
         ValueTypeOperator,
         ValueTypeOperand,
         memorySpace,
@@ -1046,11 +595,11 @@ namespace dftefe
               typename ValueTypeOperand,
               utils::MemorySpace memorySpace,
               size_type          dim>
-    EFEOverlapOperatorContext<ValueTypeOperator,
-                              ValueTypeOperand,
-                              memorySpace,
-                              dim>::
-      EFEOverlapOperatorContext(
+    PristineEFEOverlapOperatorContext<ValueTypeOperator,
+                                      ValueTypeOperand,
+                                      memorySpace,
+                                      dim>::
+      PristineEFEOverlapOperatorContext(
         const FEBasisManager<ValueTypeOperand,
                              ValueTypeOperator,
                              memorySpace,
@@ -1076,7 +625,7 @@ namespace dftefe
 
       std::shared_ptr<utils::MemoryStorage<ValueTypeOperator, memorySpace>>
         basisOverlap;
-      EFEOverlapOperatorContextInternal::computeBasisOverlapMatrix<
+      PristineEFEOverlapOperatorContextInternal::computeBasisOverlapMatrix<
         ValueTypeOperator,
         ValueTypeOperand,
         memorySpace,
@@ -1092,57 +641,8 @@ namespace dftefe
               typename ValueTypeOperand,
               utils::MemorySpace memorySpace,
               size_type          dim>
-    EFEOverlapOperatorContext<ValueTypeOperator,
-                              ValueTypeOperand,
-                              memorySpace,
-                              dim>::
-      EFEOverlapOperatorContext(
-        const FEBasisManager<ValueTypeOperand,
-                             ValueTypeOperator,
-                             memorySpace,
-                             dim> &feBasisManagerX,
-        const FEBasisManager<ValueTypeOperand,
-                             ValueTypeOperator,
-                             memorySpace,
-                             dim> &feBasisManagerY,
-        const FEBasisDataStorage<ValueTypeOperator, memorySpace>
-          &classicalBlockBasisDataStorage,
-        const FEBasisDataStorage<ValueTypeOperator, memorySpace>
-          &enrichmentBlockEnrichmentBasisDataStorage,
-        const FEBasisDataStorage<ValueTypeOperator, memorySpace>
-          &             enrichmentBlockClassicalBasisDataStorage,
-        const size_type maxCellTimesNumVecs)
-      : d_feBasisManagerX(&feBasisManagerX)
-      , d_feBasisManagerY(&feBasisManagerY)
-      , d_maxCellTimesNumVecs(maxCellTimesNumVecs)
-      , d_cellStartIdsBasisOverlap(0)
-    {
-      utils::throwException(
-        &(feBasisManagerX.getBasisDofHandler()) ==
-          &(feBasisManagerY.getBasisDofHandler()),
-        "feBasisManager of X and Y vectors are not from same basisDofhandler");
-
-      std::shared_ptr<utils::MemoryStorage<ValueTypeOperator, memorySpace>>
-        basisOverlap;
-      EFEOverlapOperatorContextInternal::computeBasisOverlapMatrix<
-        ValueTypeOperator,
-        ValueTypeOperand,
-        memorySpace,
-        dim>(classicalBlockBasisDataStorage,
-             enrichmentBlockEnrichmentBasisDataStorage,
-             enrichmentBlockClassicalBasisDataStorage,
-             basisOverlap,
-             d_cellStartIdsBasisOverlap,
-             d_dofsInCell);
-      d_basisOverlap = basisOverlap;
-    }
-
-    template <typename ValueTypeOperator,
-              typename ValueTypeOperand,
-              utils::MemorySpace memorySpace,
-              size_type          dim>
     void
-    EFEOverlapOperatorContext<
+    PristineEFEOverlapOperatorContext<
       ValueTypeOperator,
       ValueTypeOperand,
       memorySpace,
@@ -1192,7 +692,7 @@ namespace dftefe
       // perform Ax on the local part of A and x
       // (A = discrete Overlap operator)
       //
-      EFEOverlapOperatorContextInternal::computeAxCellWiseLocal(
+      PristineEFEOverlapOperatorContextInternal::computeAxCellWiseLocal(
         basisOverlapInAllCells,
         X.begin(),
         Y.begin(),
@@ -1219,10 +719,10 @@ namespace dftefe
               utils::MemorySpace memorySpace,
               size_type          dim>
     const utils::MemoryStorage<ValueTypeOperator, memorySpace> &
-    EFEOverlapOperatorContext<ValueTypeOperator,
-                              ValueTypeOperand,
-                              memorySpace,
-                              dim>::getBasisOverlapInAllCells() const
+    PristineEFEOverlapOperatorContext<ValueTypeOperator,
+                                      ValueTypeOperand,
+                                      memorySpace,
+                                      dim>::getBasisOverlapInAllCells() const
     {
       return *(d_basisOverlap);
     }
@@ -1232,11 +732,11 @@ namespace dftefe
               utils::MemorySpace memorySpace,
               size_type          dim>
     utils::MemoryStorage<ValueTypeOperator, memorySpace>
-    EFEOverlapOperatorContext<ValueTypeOperator,
-                              ValueTypeOperand,
-                              memorySpace,
-                              dim>::getBasisOverlapInCell(const size_type
-                                                            cellId) const
+    PristineEFEOverlapOperatorContext<
+      ValueTypeOperator,
+      ValueTypeOperand,
+      memorySpace,
+      dim>::getBasisOverlapInCell(const size_type cellId) const
     {
       std::shared_ptr<utils::MemoryStorage<ValueTypeOperator, memorySpace>>
                       basisOverlapStorage = d_basisOverlap;
@@ -1255,13 +755,13 @@ namespace dftefe
               utils::MemorySpace memorySpace,
               size_type          dim>
     utils::MemoryStorage<ValueTypeOperator, memorySpace>
-    EFEOverlapOperatorContext<ValueTypeOperator,
-                              ValueTypeOperand,
-                              memorySpace,
-                              dim>::getBasisOverlap(const size_type cellId,
-                                                    const size_type basisId1,
-                                                    const size_type basisId2)
-      const
+    PristineEFEOverlapOperatorContext<
+      ValueTypeOperator,
+      ValueTypeOperand,
+      memorySpace,
+      dim>::getBasisOverlap(const size_type cellId,
+                            const size_type basisId1,
+                            const size_type basisId2) const
     {
       std::shared_ptr<utils::MemoryStorage<ValueTypeOperator, memorySpace>>
         basisOverlapStorage = d_basisOverlap;
